backend/ft_trans/src/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.8
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
redis==5.0.3
rest-framework-simplejwt==0.0.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
twisted==24.3.0
txaio==23.1.1
typing-extensions==4.10.0
zope.interface==6.2
requests
python-dotenv